create test class and cleanup
